# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: FlowBook API
  description: |
    FlowBook — Booking & Scheduling Platform API.
    Go Echo + pgx + sqlc + Supabase pooler 6543.
    Store UTC (timestamptz), render in organization.timezone (default Asia/Jakarta).
    Calendar engine cached 30s, EXCLUDE anti double-book.
  version: 1.0.0
  contact:
    name: FlowBook
    url: https://flowbook.example.com
servers:
  - url: https://flowbook-api-xxx.koyeb.app/api/v1
    description: Production (Koyeb)
  - url: http://localhost:8080/api/v1
    description: Local
...
security: []
paths:
...
  /services:
    get:
      tags: [Services]
      summary: List active services (public)
      operationId: listServices
      parameters:
        - in: query
          name: organizationId
          schema:
            type: string
            format: uuid
          description: Filter by org (optional, defaults to request org)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
    post:
      tags: [Services]
      summary: Create service (OWNER)
      operationId: createService
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /dashboard:
    get:
      tags: [Dashboard]
      summary: Dashboard aggregates — 5-row revenue, bookings, occupancy with DATE_TRUNC + GROUP BY in DB (idx_bookings_start_at)
      description: |
        5-row dashboard: kpi {revenue, bookings, occupancy, avgTicket, delta}, area 10 titik Nov2025->Agu2026 (DATE_TRUNC month), pie Classic Cut 35%, bar Andi 90/Bayu70/Sari20, heatmap 7x15, topCustomers 15 Siti 18x, recent 10, insights {busiestMonth Des 2025, cancelRate 7.2%, utilization}. Queries use tstzrange + DATE_TRUNC + GROUP BY in DB with index start_at, not JS. OWNER full, STAFF scoped miliknya.
      operationId: getDashboard
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date
            example: '2025-11-01'
          description: Start date inclusive YYYY-MM-DD, interpreted in tz
        - in: query
          name: to
          schema:
            type: string
            format: date
            example: '2026-08-24'
          description: End date inclusive YYYY-MM-DD, interpreted in tz
        - in: query
          name: granularity
          schema:
            type: string
            enum: [day, week, month]
            default: month
          description: Bucket for revenueSeries — DATE_TRUNC granularity in DB
        - in: query
          name: tz
          schema:
            type: string
            default: Asia/Jakarta
            example: Asia/Jakarta
          description: IANA timezone — bucketing via AT TIME ZONE in DB, render tabular-nums
      responses:
        '200':
          description: Aggregated dashboard data — 5 rows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /ws:
    get:
      tags: [Realtime]
      summary: WebSocket upgrade — subscribes to org slot_taken broadcast
      operationId: wsUpgrade
      parameters:
        - in: query
          name: orgId
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '101':
          description: Switching Protocols
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token 15m - Authorization Bearer token"
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden — role insufficient
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationError:
      description: Validation error (422 Zod-compatible)
      content:
        application/json:
          schema:
            type: object
            required: [error, details]
            properties:
              error:
                type: string
                example: validation_error
              message:
                type: string
                example: Validation failed
              details:
                type: array
                items:
                  type: object
                  required: [field, message]
                  properties:
                    field:
                      type: string
                      example: email
                    message:
                      type: string
                      example: Invalid email
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
...
    Service:
      type: object
      required: [id, organizationId, name, durationMinutes, priceCents]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
          example: Classic Cut
        description:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          example: 30
        bufferMinutes:
          type: integer
          example: 10
        priceCents:
          type: integer
          example: 85000
        color:
          type: string
          example: '#7c3aed'
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateServiceRequest:
      type: object
      required: [name, durationMinutes, priceCents]
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
        durationMinutes:
          type: integer
          minimum: 5
          maximum: 480
        bufferMinutes:
          type: integer
          minimum: 0
          default: 10
        priceCents:
          type: integer
          minimum: 0
        color:
          type: string
        isActive:
          type: boolean
          default: true
...
    Booking:
      type: object
      required: [id, organizationId, serviceId, staffId, customerName, customerEmail, startAt, endAt, status]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        staffId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
          nullable: true
        customerName:
          type: string
        customerEmail:
          type: string
          format: email
        customerPhone:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED, NO_SHOW]
        paymentStatus:
          type: string
          enum: [UNPAID, PAID, REFUNDED]
        stripeSessionId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    DashboardResponse:
      type: object
      required: [kpi, revenueSeries]
      properties:
        kpi:
          type: object
          required: [totalBookings, confirmedBookings, totalRevenueCents, avgTicketCents, occupancyPct]
          properties:
            totalBookings:
              type: integer
              description: Total bookings in range — tabular-nums
            confirmedBookings:
              type: integer
            cancelledBookings:
              type: integer
            totalRevenueCents:
              type: integer
              description: Sum price_cents CONFIRMED/COMPLETED — Rp tabular-nums
            avgTicketCents:
              type: integer
            occupancyPct:
              type: number
              description: CONFIRMED/COMPLETED / total *100
            deltaRevenuePct:
              type: number
              nullable: true
              description: vs previous period equal length
            deltaBookingsPct:
              type: number
              nullable: true
        revenueSeries:
          type: array
          description: Area 10 titik Nov2025->Agu2026 — DATE_TRUNC + GROUP BY in DB with index start_at
          items:
            type: object
            required: [period, revenueCents, bookingsCount]
            properties:
              period:
                type: string
                format: date-time
                description: Bucket start in target tz, ISO8601 UTC
              revenueCents:
                type: integer
              bookingsCount:
                type: integer
              label:
                type: string
                nullable: true
        topServices:
          type: array
          description: Pie Classic Cut 35% — share of bookings
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              color:
                type: string
              bookingsCount:
                type: integer
              revenueCents:
                type: integer
              percentage:
                type: number
                description: Share of confirmed bookings
        bookingsByStaff:
          type: array
          description: Bar Andi 90/Bayu 70/Sari 20 — GROUP BY staff.id, DATE_TRUNC in DB
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              bookingsCount:
                type: integer
              revenueCents:
                type: integer
        bookingsByHour:
          type: array
          description: Hourly distribution — EXTRACT(HOUR AT TIME ZONE tz) + GROUP BY in DB
          items:
            type: object
            properties:
              hour:
                type: integer
              bookingsCount:
                type: integer
        heatmap:
          type: array
          description: Heatmap 7x15 Jam Sibuk — DOW 0-6 x Hour 7-21, GROUP BY dow, hour in DB with AT TIME ZONE
          items:
            type: object
            required: [dow, hour, count]
            properties:
              dow:
                type: integer
                minimum: 0
                maximum: 6
              hour:
                type: integer
                minimum: 0
                maximum: 23
              count:
                type: integer
        topCustomers:
          type: array
          description: Top 15 Loyal — Siti 18x, GROUP BY customer_email in DB
          items:
            type: object
            properties:
              customerEmail:
                type: string
              customerName:
                type: string
              bookingsCount:
                type: integer
              totalSpentCents:
                type: integer
              lastBookingAt:
                type: string
                format: date-time
                nullable: true
        recentBookings:
          type: array
          description: Recent 10 bookings with status Badge
          items:
            $ref: '#/components/schemas/Booking'
        insights:
          type: object
          description: Insight row — busiestMonth Des 2025, cancelRate 7.2%, utilization
          properties:
            busiestMonth:
              type: string
              example: "Des 2025"
            busiestMonthCount:
              type: integer
            busiestMonthRevenue:
              type: integer
            cancelRate:
              type: number
              example: 7.2
            utilization:
              type: number
              example: 68